Rename: layout-without-chat 제거 후 routes 구조 원상복구
